+++ webapi/src/it/scala/org/knora/webapi/ITKnoraLiveSpec.scala
@@ -15,7 +15,6 @@ import org.scalatest.matchers.should.Matchers
 import org.scalatest.wordspec.AnyWordSpec
 import spray.json._
 import zio._
-import zio.logging.backend.SLF4J
 
 import java.util.concurrent.TimeUnit
 import scala.concurrent.Await
@@ -65,11 +64,7 @@ abstract class ITKnoraLiveSpec
    * `Bootstrap` will ensure that everything is instantiated when the Runtime is created
    * and cleaned up when the Runtime is shutdown.
    */
-  private val bootstrap: ZLayer[
-    Any,
-    Any,
-    Environment
-  ] = ZLayer.empty ++ Runtime.removeDefaultLoggers ++ SLF4J.slf4j ++ effectLayers
+  private val bootstrap = util.Logger.textLogger() >>> effectLayers
 
   // create a configured runtime
   val runtime = Unsafe.unsafe { implicit u =>
